+++ lib/elixir/test/elixir/exception_test.exs
@@ -6,11 +6,39 @@ defmodule Kernel.ExceptionTest do
   # Ensure fields passed through an expression are valid
   defexception Custom, ~w(message)a
 
+  test "raise preserves the stacktrace" do
+    stacktrace =
+      try do
+        raise "a"
+      rescue _ ->
+        [top|_] = System.stacktrace
+        top
+      end
+    file = __ENV__.file |> Path.relative_to_cwd |> List.from_char_data!
+    assert {Kernel.ExceptionTest, :"test raise preserves the stacktrace", _,
+           [file: ^file, line: 12]} = stacktrace
+  end
+
   test "is_exception" do
     assert is_exception(RuntimeError.new)
     refute is_exception(empty_tuple)
     refute is_exception(a_tuple)
     refute is_exception(a_list)
+  end
+
+  test "message" do
+    defmodule BadException do
+      def message(_) do
+        raise "oops"
+      end
+    end
+
+    message = "Got RuntimeError with message \"oops\" " <>
+              "while retrieving message for {Kernel.ExceptionTest.BadException}"
+
+    assert_raise ArgumentError, message, fn ->
+      Exception.message({BadException})
+    end
   end
 
   test "normalize" do
@@ -161,49 +189,36 @@ defmodule Kernel.ExceptionTest do
            ~r"#Function<\d\.\d+/0 in Kernel\.ExceptionTest\.test format_fa/1>/1"
   end
 
+  import Exception, only: [message: 1]
+
   test "runtime error message" do
-    assert RuntimeError.new.message == "runtime error"
-    assert RuntimeError.new(message: "exception").message == "exception"
+    assert RuntimeError.new |> message == "runtime error"
+    assert RuntimeError.new(message: "exception") |> message == "exception"
   end
 
   test "argument error message" do
-    assert ArgumentError.new.message == "argument error"
-    assert ArgumentError.new(message: "exception").message == "exception"
+    assert ArgumentError.new |> message == "argument error"
+    assert ArgumentError.new(message: "exception") |> message == "exception"
   end
 
   test "undefined function message" do
-    assert UndefinedFunctionError.new.message == "undefined function"
-    assert UndefinedFunctionError.new(module: Foo, function: :bar, arity: 1).message == "undefined function: Foo.bar/1"
-    assert UndefinedFunctionError.new(module: nil, function: :bar, arity: 0).message == "undefined function: nil.bar/0"
+    assert UndefinedFunctionError.new |> message == "undefined function"
+    assert UndefinedFunctionError.new(module: Foo, function: :bar, arity: 1) |> message ==
+           "undefined function: Foo.bar/1"
+    assert UndefinedFunctionError.new(module: nil, function: :bar, arity: 0) |> message ==
+           "undefined function: nil.bar/0"
   end
 
   test "function clause message" do
-    assert FunctionClauseError.new.message == "no function clause matches"
-    assert FunctionClauseError.new(module: Foo, function: :bar, arity: 1).message == "no function clause matching in Foo.bar/1"
+    assert FunctionClauseError.new |> message ==
+           "no function clause matches"
+    assert FunctionClauseError.new(module: Foo, function: :bar, arity: 1) |> message ==
+           "no function clause matching in Foo.bar/1"
   end
 
   test "erlang error message" do
-    assert ErlangError.new(original: :sample).message == "erlang error: :sample"
-  end
-
-  test "raise preserves the stacktrace" do
-    stacktrace =
-      try do
-        raise "a"
-      rescue _ ->
-        [top|_] = System.stacktrace
-        top
-      end
-    file = __ENV__.file |> Path.relative_to_cwd |> List.from_char_data!
-    assert {Kernel.ExceptionTest, :"test raise preserves the stacktrace", _,
-           [file: ^file, line: 192]} = stacktrace
-  end
-
-  test "defexception" do
-    defexception SampleError, message: nil do
-      # Check do block is properly inline
-      def exception(_), do: SampleError[message: "hello"]
-    end
+    assert ErlangError.new(original: :sample) |> message ==
+           "erlang error: :sample"
   end
 
   defp empty_tuple, do: {}